# swagger.json
{"swagger":"2.0","info":{"description":"CMPasS project documentation","version":"0.4.0","title":"cmpaas-nestjs"},"basePath":"/v1","tags":[{"name":"CMPaaS","description":""}],"schemes":["http","https"],"securityDefinitions":{"bearer":{"type":"apiKey","name":"Authorization","in":"header"}},"paths":{"/":{"get":{"responses":{"200":{"description":""}},"produces":["application/json"],"consumes":["application/json"]}},"/users":{"post":{"parameters":[{"name":"CreateUserDto","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/CreateUserDto"}}],"responses":{"200":{"description":"New user was created."},"404":{"description":"New user was not created."},"503":{"description":"Server error."}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]},"get":{"responses":{"200":{"description":"User was find."},"404":{"description":"User was not found."},"503":{"description":"Server error."}},"tags":["users"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/users/email/{email}":{"get":{"parameters":[{"name":"email","required":true,"in":"path","description":"check email","type":"string"}],"responses":{"200":{"description":"Email was found."},"404":{"description":"Email wasn t found."},"503":{"description":"Server error."}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]}},"/users/username/{username}":{"get":{"parameters":[{"name":"username","required":true,"in":"path","description":"check username","type":"string"}],"responses":{"200":{"description":"Username was found."},"404":{"description":"Username wasn t found."},"503":{"description":"Server error."}},"tags":["users"],"produces":["application/json"],"consumes":["application/json"]}},"/auth/login":{"post":{"parameters":[{"name":"body","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/UserLoginDto"}}],"responses":{"200":{"description":"Login aproved."},"404":{"description":"Login was not approved."},"503":{"description":"Server error."}},"tags":["auth"],"produces":["application/json"],"consumes":["application/json"]}},"/auth/reset/password":{"post":{"parameters":[{"name":"body","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/ResetLoginDto"}}],"responses":{"200":{"description":"Email aproved."},"404":{"description":"Email was not approved."},"503":{"description":"Server error."}},"tags":["auth"],"produces":["application/json"],"consumes":["application/json"]}},"/auth/new/password":{"post":{"parameters":[{"name":"body","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/UserLoginDto"}}],"responses":{"200":{"description":"Password was changed."},"404":{"description":"Password was not approved."},"503":{"description":"Server error."}},"tags":["auth"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/maps":{"post":{"parameters":[{"name":"CreateMapDto","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/CreateMapDto"}}],"responses":{"200":{"description":"Map was created."},"404":{"description":"Map was not created."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"get":{"responses":{"200":{"description":"Map was find."},"404":{"description":"Map was not find."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/maps/export/cmaps":{"post":{"parameters":[{"name":"body","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/ExportMapDto"}}],"responses":{"200":{"description":"Map export has successful"},"404":{"description":"Map export has not successful"},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/maps/{idmap}/versions":{"get":{"parameters":[{"name":"idmap","required":true,"in":"path","description":"enter map ID","type":"string"}],"responses":{"200":{"description":"Map version was find."},"404":{"description":"Map version was not find."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/maps/{idmap}/versions/{idversion}":{"get":{"parameters":[{"name":"idversion","required":true,"in":"path","description":"enter version ID","type":"string"},{"name":"idmap","required":true,"in":"path","description":"enter map ID","type":"string"}],"responses":{"200":{"description":"Map version was find."},"404":{"description":"Map version was not find."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"parameters":[{"name":"idversion","required":true,"in":"path","description":"enter version ID","type":"string"},{"name":"idmap","required":true,"in":"path","description":"enter map ID","type":"string"}],"responses":{"200":{"description":"Version was deleted."},"404":{"description":"version was not deleted."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/maps/{idmap}":{"put":{"parameters":[{"name":"body","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/EditMapDto"}},{"name":"idmap","required":true,"in":"path","description":"enter map ID","type":"string"}],"responses":{"200":{"description":"Map was updated."},"404":{"description":"Map was not updated."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]},"delete":{"parameters":[{"name":"idmap","required":true,"in":"path","description":"enter map ID","type":"string"}],"responses":{"200":{"description":"Map was deleted."},"404":{"description":"Map was not deleted."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}},"/maps/{idmap}/content":{"post":{"parameters":[{"name":"body","required":true,"in":"body","isArray":false,"schema":{"$ref":"#/definitions/NewMapDto"}},{"name":"idmap","required":true,"in":"path","description":"enter map ID","type":"string"}],"responses":{"200":{"description":"New content was created."},"404":{"description":"New content was not created."},"503":{"description":"Server error."}},"tags":["maps"],"security":[{"bearer":[]}],"produces":["application/json"],"consumes":["application/json"]}}},"definitions":{"CreateUserDto":{"type":"object","properties":{"firstname":{"type":"string"},"surname":{"type":"string"},"username":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["firstname","surname","username","email","password"]},"UserLoginDto":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}},"required":["username","password"]},"ResetLoginDto":{"type":"object","properties":{"email":{"type":"string"}},"required":["email"]},"CreateMapDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"question":{"type":"string"},"keywords":{"type":"array","items":{"type":"string"}},"content":{"type":"object"}},"required":["title","description","question","keywords","content"]},"nodeDataArrayDto":{"type":"object","properties":{"text":{"type":"string"},"loc":{"type":"string"},"category":{"type":"string"},"error":{"type":"string"},"key":{"type":"number"}},"required":["text","loc","category","error","key"]},"linkDataArrayDto":{"type":"object","properties":{"from":{"type":"number"},"to":{"type":"number"},"category":{"type":"string"},"error":{"type":"string"},"points":{"type":"array","items":{"type":"string"}}},"required":["from","to","category","error","points"]},"ExportMapDto":{"type":"object","properties":{"class":{"type":"string"},"nodeDataArray":{"$ref":"#/definitions/nodeDataArrayDto"},"linkDataArray":{"$ref":"#/definitions/linkDataArrayDto"}},"required":["class","nodeDataArray","linkDataArray"]},"EditMapDto":{"type":"object","properties":{"_id":{"type":"string"},"title":{"type":"string"},"description":{"type":"string"},"question":{"type":"string"},"keywords":{"type":"array","items":{"type":"string"}}},"required":["_id","title","description","question","keywords"]},"NewMapDto":{"type":"object","properties":{"_id":{"type":"string"},"content":{"type":"object"}},"required":["_id","content"]}}}